added filters for tags

+++ GAM250/Assets/LineUp/Scripts/MovmentData.cs
@@ -6,7 +6,7 @@ namespace LineUp
     [System.Serializable]
     public class MovmentData
     {
-        public int id;
+        public float id;
         public Transform transformToTrack;
         public string movementString;
         public Vector3 lastRecordedPosition;
